+++ rnalysis/utils/enrichment_runner.py
@@ -8,6 +8,7 @@ from pathlib import Path
 from typing import Iterable, List, Tuple, Union, Collection, Set, Dict
 
 import joblib
+import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -19,11 +20,12 @@ from statsmodels.stats.descriptivestats import sign_test
 from tqdm.auto import tqdm
 from typing_extensions import Literal
 
-from rnalysis.utils import ontology, io, parsing, settings, validation, generic
+from rnalysis.utils import ontology, io, parsing, settings, validation, generic, param_typing
 from rnalysis.utils.param_typing import PARALLEL_BACKENDS, GRAPHVIZ_FORMATS
 
 try:
     import xlmhg
+
     xlmhg.test.np.float = np.float64
 
     logging.getLogger('xlmhg').setLevel(50)  # suppress warnings from xlmhg module
@@ -52,6 +54,29 @@ def does_python_version_support_single_set():
 
 
 XLMHG_SUPPORTED = does_python_version_support_single_set()
+
+
+class Size:
+    def __init__(self, size: int):
+        self.size = size
+
+
+def dot_scaling_func(values: Union[float, np.ndarray, pd.Series]):
+    return np.sqrt(values) * 12
+
+
+class sizeHandler:
+    def __init__(self, size: int):
+        self.size = size
+
+    def legend_artist(self, legend, orig_handle, fontsize, handlebox):
+        x0, y0 = handlebox.xdescent, handlebox.ydescent
+        radius = np.sqrt((dot_scaling_func(self.size)) / np.pi)
+        patch = mpatches.Circle((x0, y0), radius, facecolor='black',
+                                edgecolor='black', hatch='xx', lw=3,
+                                transform=handlebox.get_transform())
+        handlebox.add_artist(patch)
+        return patch
 
 
 class EnrichmentRunner:
@@ -86,7 +111,8 @@ class EnrichmentRunner:
                                  'pre-sorted and ranked by the user'}
     printout_params = "appear in the Attribute Reference Table"
 
-    def __init__(self, genes: Union[set, np.ndarray], attributes: Union[Iterable, str, int], alpha: float,
+    def __init__(self, genes: Union[set, np.ndarray], attributes: Union[Iterable, str, int],
+                 alpha: param_typing.Fraction,
                  attr_ref_path: str, return_nonsignificant: bool, save_csv: bool, fname: str, return_fig: bool,
                  plot_horizontal: bool, set_name: str,
                  parallel_backend: Literal[PARALLEL_BACKENDS], enrichment_func_name: str, biotypes=None,
@@ -233,9 +259,9 @@ class EnrichmentRunner:
                 return self._xlmhg_enrichment
             else:
                 if not HAS_XLMHG:
-                    warnings.warn(f"Package 'xlmhg' is not installed. \n"
-                                  f"If you want to run single-set enrichment analysis, "
-                                  f"please install package 'xlmhg' and try again. ")
+                    warnings.warn("Package 'xlmhg' is not installed. \n"
+                                  "If you want to run single-set enrichment analysis, "
+                                  "please install package 'xlmhg' and try again. ")
                 if not XLMHG_SUPPORTED:
                     python_version = sys.version_info
                     warnings.warn(f"Your version of Python ({python_version[0]}.{python_version[1]}) "
@@ -518,28 +544,39 @@ class EnrichmentRunner:
         return self.enrichment_bar_plot(ylabel=self.ENRICHMENT_SCORE_YLABEL,
                                         title=f"Enrichment for gene set '{self.set_name}'")
 
-    def enrichment_bar_plot(self, n_bars: int = 'all', center_bars: bool = True,
-                            ylabel: str = r"$\log_2$(Fold Enrichment)",
-                            title: str = 'Enrichment results',
-                            ylim: Union[float, Literal['auto']] = 'auto') -> plt.Figure:
+    def enrichment_bar_plot(self, n_bars: int = 'all', center_bars: bool = True, show_expected=False,
+                            ylabel: str = r"$\log_2$(Fold Enrichment)", title: str = 'Enrichment results',
+                            ylim: Union[float, Literal['auto']] = 'auto',
+                            plot_style: Literal['bar', 'lollipop'] = 'bar') -> plt.Figure:
 
         """
         Receives a DataFrame output from an enrichment function and plots it in a bar plot. \
         For the clarity of display, complete depletion (linear enrichment = 0) \
         appears with the smallest value in the scale.
 
-        :param title:
-        :type title:
-        :param n_bars:
-        :type n_bars:
-        :param ylabel: plot ylabel.
-        :type ylabel: str
-        :param center_bars: if True, centers the bars around Y=0. Otherwise, ylim is determined by min/max values.
-        :type center_bars: bool (default True)
+        :param n_bars: number of bars to display in the bar plot. If n_bars='all', \
+         all the results will be displayed on the graph. Otherwise, only the top n results will be displayed on the graph.
+        :type n_bars: int > 1 or 'all' (default='all')
+        :param title: plot title.
+        :type title: str
+        :param ylabel: plot y-axis label.
+        :type ylabel: str (default=r"$\log_2$(Fold Enrichment)")
+        :param center_bars: if True, center the bars around Y=0. Otherwise, ylim is determined by min/max values.
+        :type center_bars: bool (default=True)
+        :param ylim: set the Y-axis limits. If `ylim`='auto', determines the axis limits automatically based on the data. \
+        If `ylim` is a number, set the Y-axis limits to [-ylim, ylim].
+        :type ylim: float or 'auto' (default='auto')
+        :param plot_style: style for the plot. Either 'bar' for a bar plot or 'lollipop' for a lollipop plot \
+        in which the lollipop size indicates the size of the observed gene set.
+        :type plot_style: 'bar' or 'lollipop' (default='bar')
+        :param show_expected: if True, the observed/expected values will be shown on the plot.
+        :type show_expected: bool (default=False)
         :return: Figure object containing the bar plot
         :rtype: matplotlib.figure.Figure instance
         """
-        plt.style.use('seaborn-white')
+        assert plot_style in ['bar', 'lollipop'], \
+            f"'plot_style' must be 'bar' or 'lollipop', instaed got '{plot_style}'."
+
         # determine number of entries/bars to plot
         if n_bars != 'all':
             assert isinstance(n_bars, int)
@@ -559,18 +596,20 @@ class EnrichmentRunner:
         # choose functions and parameters according to the graph's orientation (horizontal vs vertical)
         if self.plot_horizontal:
             figsize = [10.5, 0.4 * (4.8 + self.results.shape[0])]
-            bar_func = plt.Axes.barh
+            bar_func = plt.Axes.barh if plot_style == 'bar' else plt.Axes.hlines
             line_func = plt.Axes.axvline
+
             cbar_location = 'bottom'
             cbar_orientation = 'horizontal'
             tick_func = plt.Axes.set_yticks
             ticklabels_func = plt.Axes.set_yticklabels
             ticklabels_kwargs = dict(fontsize=13, rotation=0)
+
             for lst in (enrichment_names, enrichment_scores, enrichment_pvalue):
                 lst.reverse()
         else:
             figsize = [0.5 * (4.8 + self.results.shape[0]), 4.2]
-            bar_func = plt.Axes.bar
+            bar_func = plt.Axes.bar if plot_style == 'bar' else plt.Axes.vlines
             line_func = plt.Axes.axhline
             cbar_location = 'left'
             cbar_orientation = 'vertical'
@@ -605,9 +644,24 @@ class EnrichmentRunner:
 
         # generate bar plot
         fig, ax = plt.subplots(tight_layout=True, figsize=figsize)
-        bar = bar_func(ax, range(len(enrichment_names)), enrichment_scores, color=colors, edgecolor='black',
-                       linewidth=1, zorder=2)
-        bar.tick_labels = enrichment_names
+
+        args = (ax, range(len(enrichment_names)), enrichment_scores) if plot_style == 'bar' else (
+            ax, range(len(enrichment_names)), 0, enrichment_scores)
+        kwargs = dict(linewidth=1, edgecolor='black') if plot_style == 'bar' else dict(linewidth=5)
+
+        graph = bar_func(*args, color=colors, zorder=2, **kwargs)
+        graph.tick_labels = enrichment_names
+
+        if plot_style == 'lollipop':
+            x, y = (enrichment_scores, range(len(enrichment_names))) if self.plot_horizontal else (range(
+                len(enrichment_names)), enrichment_scores)
+            ax.scatter(x, y, color=colors, zorder=3, s=dot_scaling_func(results['obs']))
+
+            legend_sizes = [Size(i) for i in [1, 10, 100, 1000]]
+            ax.legend(legend_sizes, [f"observed={i.size}" for i in legend_sizes],
+                      labelspacing=1.5, borderpad=1.5, draggable=True,
+                      handler_map={size: sizeHandler(size.size) for size in legend_sizes}, loc='lower right')
+
         # determine bounds, and enlarge the bound by a small margin (0.5%) so nothing gets cut out of the figure
         if ylim == 'auto':
             bounds = np.array([np.floor(-max_score), np.ceil(max_score)]) * 1.005
@@ -626,23 +680,31 @@ class EnrichmentRunner:
         # title
         ax.set_title(title, fontsize=18)
         # add significance asterisks
-        for col, sig in zip(bar, enrichment_pvalue):
+        for i, (score, sig) in enumerate(zip(enrichment_scores, enrichment_pvalue)):
             asterisks, fontweight = self._get_pval_asterisk(sig, self.alpha)
             if self.plot_horizontal:
-                x = col._width
-                y = col.xy[1] + 0.5 * col._height
-                valign = 'center'
-                halign = 'left' if np.sign(col._width) == 1 else 'right'
-                rotation = 270 if np.sign(col._width) == 1 else 90
+                x = score
+                y = i
+                valign = 'bottom'
+                halign = 'center'
+                rotation = 270 if np.sign(score) == 1 else 90
+                rotation_mode = 'anchor'
             else:
-                x = col.xy[0] + 0.5 * col._width
-                y = col._height
-                valign = 'bottom' if np.sign(col._height) == 1 else 'top'
+                x = i
+                y = score
+                valign = 'bottom' if np.sign(score) == 1 else 'top'
                 halign = 'center'
                 rotation = 0
+                rotation_mode = 'default'
 
-            ax.text(x=x, y=y, s=asterisks, fontname='DejaVu Sans', fontweight=fontweight, rotation=rotation,
-                    fontsize=9, horizontalalignment=halign, verticalalignment=valign, zorder=1)
+            if show_expected:
+                extra_text = f"{results['obs'].iloc[i]}/{results['exp'].iloc[i]:.1f}"
+                asterisks = f"{extra_text}\n{asterisks}" if self.plot_horizontal or np.sign(
+                    score) == 1 else f"{asterisks}\n{extra_text}"
+            ax.text(x=x, y=y, s=asterisks, fontname='DejaVu Sans',
+                    fontweight=fontweight, rotation=rotation, fontsize=9,
+                    horizontalalignment=halign, verticalalignment=valign, zorder=3, rotation_mode=rotation_mode)
+
         # despine
         _ = [ax.spines[side].set_visible(False) for side in ['top', 'bottom', 'left', 'right']]
         # center bars